requirements.txt
begins
dspy-ai
python-dotenv
openai
google-generativeai
anthropic
textgrad
python-slugify